Use LD_LIBRARY_PATH on linux (#18)
On some systems LD_LIBRARY_PATH is set to help find libraries. When
present this can help find the required library without requiring the
user to explicitly set it. This is specifically interesting for cases
like ubuntu snaps where libraries are installed to non-standard paths and
LD_LIBRARY_PATH is set to find them.